FIX allocate suitable stack for tasks FIX give gyroscope a init value 0 to avoid fault ADD nomalization the quaternion

+++ Drivers/Peripherals/Inc/mpu6050.h
@@ -17,6 +17,6 @@ void MPU6050_Read_Scaler(uint8_t* buf,uint8_t* pos,uint8_t size);
 /**
  * @brief read Gyroscope data from mpu6050
 */
-uint8_t MPU6050_Get_Gyroscope(float* alpha,float* beta,float* gama);
+uint8_t MPU6050_Get_Gyroscope(double* alpha,double* beta,double* gama);
 
 #endif
