+++ Drivers/Peripherals/Src/mpu6050.c
@@ -62,7 +62,7 @@ void MPU6050_Read_Scaler(uint8_t* buf,uint8_t* pos,uint8_t size)
 
 }
 
-uint8_t MPU6050_Get_Gyroscope(float* alpha,float* beta,float* gama)
+uint8_t MPU6050_Get_Gyroscope(double* alpha,double* beta,double* gama)
 {
     short xout,yout,zout;
     uint8_t buffer[6] = {0};
